File headers that `package-install-file' likes.
Goal is to successfully install this package via `package-install-file'.
Desire is to be able to install this package via MELPA. In order to do
so, reformatted the package definition content into the current format
for EMACS.

Tested it on GNU Emacs 24.5.1 (x86_64-apple-darwin15.4.0, NS
apple-appkit-1404.46) of 2016-06-08 on vela and it worked correctly; the
package installed.